Add files via upload
Random Forest is an ensemble learning method used primarily for classification and regression tasks in machine learning. It combines multiple decision trees to create a more robust and accurate model. By aggregating the predictions of several decision trees, a Random Forest aims to improve accuracy and prevent overfitting, which is a common issue with individual decision trees.