Add files via upload
A Decision Tree is a supervised machine learning algorithm used for both classification and regression tasks. It works by splitting the dataset into subsets based on the values of input features, resulting in a tree-like structure where each internal node represents a decision based on a feature, and each leaf node represents an outcome (either a class label in classification or a continuous value in regression).